% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbox_nwes}
\alias{bbox_nwes}
\title{North-west European Shelf boundary box.}
\format{
sf of North-west European Shelf boundary box.
}
\source{
https://doi.org/10.5194/essd-10-109-2018
}
\usage{
bbox_nwes
}
\description{
North-west European Shelf boundary box.
}
\keyword{datasets}
