% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-api.R
\name{get_bio_oracle_t}
\alias{get_bio_oracle_t}
\title{Title}
\usage{
get_bio_oracle_t(
  obis_recs,
  layercodes = c("BO_sstmean", "BO2_tempmean_bdmean"),
  use_defaults = TRUE
)
}
\arguments{
\item{use_defaults}{}
}
\value{

}
\description{
Title
}
