% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-emodnet.R
\name{get_emodnet_info}
\alias{get_emodnet_info}
\title{Access EMODnet WFS info}
\usage{
get_emodnet_info(
  wfs_url = "https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs"
)
}
\arguments{
\item{wfs_url}{url to the EMODnet WFS service. Defaults to the EUNIS seabed habitat map wfs.}
}
\value{
WFS available feature information
}
\description{
Access EMODnet WFS info
}
\examples{
get_emodnet_info()
}
