% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-emodnet.R
\name{get_emodnet_wfs_layers}
\alias{get_emodnet_wfs_layers}
\title{Access EMODnet WFS layers}
\usage{
get_emodnet_wfs_layers(layers, service, cache_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{layers}{character vector of EMODnet seabed habitat wfs layer names.}

\item{service}{name of the EMODnet WFS service.}

\item{cache_dir}{path to the cached data directory. Defaults to the "cached-data'
folder in the root of the current project}

\item{overwrite}{whether to re-download and overwrite already downloaded layers}
}
\value{
an sf of requested features
}
\description{
Access EMODnet WFS layers
}
\examples{
\dontrun
proj_dir <- tempdir()
cache_dir <- fs::path(proj_dir, "cached-data")
get_emodnet_wfs_layers(layers = "be000142", 
                       service = "seabed_habitats_individual_habitat_map_and_model_datasets",
                      cache_dir = cache_dir)
}
