% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-api.R
\name{get_iap_gridded_t}
\alias{get_iap_gridded_t}
\title{Title}
\usage{
get_iap_gridded_t(obis_recs, use_defaults = TRUE, bespoke_path = NULL)
}
\arguments{
\item{bespoke_path}{}
}
\value{

}
\description{
Title
}
