% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-sedm.R
\name{get_sedm}
\alias{get_sedm}
\title{Extract sedmaps data for an OBIS occurence sf}
\usage{
get_sedm(obis_recs, layers)
}
\arguments{
\item{obis_recs}{OBIS occurence sf}

\item{layers}{character vector of layer names to extract}
}
\value{
a tibble of extracted layer data
}
\description{
Extract sedmaps data for an OBIS occurence sf
}
