% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-sedmaps-api.R
\name{get_sedmaps}
\alias{get_sedmaps}
\title{Extract sedmaps data for an OBIS occurence sf}
\usage{
get_sedmaps(obis_recs, layers = names(sm_rst))
}
\arguments{
\item{obis_recs}{OBIS occurence sf}

\item{layers}{character vector of layer names to extract}
}
\value{
a tibble of extracted layer data
}
\description{
Extract sedmaps data for an OBIS occurence sf
}
