% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-spdm.R
\name{get_spdm}
\alias{get_spdm}
\title{Extract \code{sdmpredictor} raster data for an OBIS occurence \code{sf}}
\usage{
get_spdm(obis_recs, layercodes, cache_dir = NULL, crs = NULL)
}
\arguments{
\item{obis_recs}{data.frame of OBIS records}

\item{layercodes}{character vector of bio oracle layer names}

\item{cache_dir}{directory to cache data into}

\item{crs}{crs of obis records. If NULL, defaults to EPSG code 4326 (WSG 84)}
}
\value{
a tibble of extracted sdmpredictor raster data, one column per layercode.
}
\description{
Extract \code{sdmpredictor} raster data for an OBIS occurence \code{sf}
}
