% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis-to-h.R
\name{get_temp_summ_by_sp}
\alias{get_temp_summ_by_sp}
\title{Title}
\usage{
get_temp_summ_by_sp(
  sp_id,
  save_all_recs = TRUE,
  check_match = TRUE,
  use_defaults = TRUE
)
}
\arguments{
\item{use_defaults}{}
}
\value{

}
\description{
Title
}
