% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_info.R
\docType{data}
\name{layer_info}
\alias{layer_info}
\title{layer metadata}
\format{
A tibble of 33 variables and 3255 observations
}
\source{

}
\usage{
layer_info
}
\description{
Attributes of all data layers available
}
\keyword{datasets}
