% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis-to-h.R
\name{match_temp_sp}
\alias{match_temp_sp}
\title{Title}
\usage{
match_temp_sp(sp_id, layers, check_match = FALSE, trim = NULL)
}
\arguments{
\item{layers}{character vector of habitat layers to match}
}
\value{

}
\description{
function to get OBIS records for a given species and match to IAP and Bio-Oracle temperatures
if save_all_recs == TRUE, this full matched dataset will be saved before summarising
function returns a summary of temperature affinity of the species
}
