% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis-to-h.R
\name{obis_match_habitat}
\alias{obis_match_habitat}
\title{Title}
\usage{
obis_match_habitat(
  sp_id,
  layers,
  check_match = FALSE,
  validate_sp_id = FALSE,
  trim = NULL
)
}
\arguments{
\item{sp_id}{validates species Aphia ID}

\item{layers}{character vector of habitat layers to match}

\item{check_match}{whether to check validity of \code{sp_id}.}

\item{trim}{whether to trim output to a particular area}
}
\value{

}
\description{
function to get OBIS records for a given species and match to IAP and Bio-Oracle temperatures
if save_all_recs == TRUE, this full matched dataset will be saved before summarising
function returns a summary of temperature affinity of the species
}
\examples{
obis_match_habitat(sp_id = 325567, 
layers = c("BO_bathymean", "BO2_phosphateltmax_bdmean", "BO2_salinitymin_ss", "rock50cm", "april", "march"))
}
