% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedm_layers}
\alias{sedm_layers}
\title{sedmaps raster stack variable names.}
\format{
A list of variable names
\describe{
\item{sed}{sediment characteristics layers}
\item{dis}{monthly sedimentary distrurbance layers}
}
}
\usage{
sedm_layers
}
\description{
sedmaps raster stack variable names.
}
\keyword{datasets}
