% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sedm_rst}
\alias{sedm_rst}
\title{North-west European Shelf sedimentary environment.}
\format{
A raster stack with variable layers
\describe{
\item{var1}{percentage compositions of mud, sand and gravel}
\item{var1}{median grain size of the whole sediment}
\item{var1}{median grain size of the sand fractions}
\item{var1}{median grain size of the gravel fractions}
\item{var1}{carbon content of sediments}
\item{var1}{nitrogen content of sediments}
\item{var1}{percentage of seabed area covered by rock}
\item{var1}{mean depth-averaged tidal velocity and wave orbital velocity at the seabed}
\item{var1}{mean depth-averaged wave orbital velocity at the seabed}
\item{var1}{maximum depth-averaged tidal velocity at the seabed}
\item{var1}{maximum depth-averaged wave orbital velocity at the seabed}
\item{jan-dec}{mean monthly natural disturbance rates (12 layers)}
...
}
}
\source{
\url{https://doi.org/10.5194/essd-10-109-2018}
}
\usage{
sedm_rst
}
\description{
A dataset from Wilson, R. J., Speirs, D. C., Sabatino, A., and Heath, M. R.:
\emph{\strong{A synthetic map of the north-west European Shelf sedimentary environment
for applications in marine science}}, Earth Syst. Sci. Data, 10, 109-130, 2018
}
\details{
0.125° by 0.125° resolution synthetic maps of continuous properties of the north-west
European sedimentary environment, extending from the Bay of Biscay to the northern limits
of the North Sea and the Faroe Islands.

The maps are a blend of gridded survey data, statistically modelled values based on
distributions of bed shear stress due to tidal currents and waves, and bathymetric
properties.
}
\keyword{datasets}
