% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{use_data_roxygen}
\alias{use_data_roxygen}
\title{Create data roxygen documentation}
\usage{
use_data_roxygen(
  dataset_name,
  dataset_title,
  description,
  source_url = "https://example.com",
  format = "tibble",
  data
)
}
\arguments{
\item{dataset_name}{character string. variable name}

\item{dataset_title}{character string. Dataset title}

\item{description}{character string. Dataset description}

\item{source_url}{character string. Dataset source URL}

\item{format}{character string. Dataset format}

\item{data}{data.frame of data being documented}
}
\value{
a format description string
}
\description{
Create data roxygen documentation
}
