% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-emodnet.R
\name{get_emodnet_wfs_layers}
\alias{get_emodnet_wfs_layers}
\title{Access EMODnet EUNIS Seabed Habitat layers}
\usage{
get_emodnet_wfs_layers(
  layers = c("be000142", "be000143", "be000144", "bg003005", "es000001", "es001000"),
  cache_dir = NULL,
  wfs_url = "https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs"
)
}
\arguments{
\item{layers}{character vector of EMODnet seabed habitat wfs layer names.}

\item{cache_dir}{path to the cached data directory. Defaults to the "cached-data'
folder in the root of the current project}

\item{wfs_url}{url to the EMODnet WFS service. Defaults to the EUNIS seabed habitat map wfs.}
}
\value{
an sf of requested features
}
\description{
Access EMODnet EUNIS Seabed Habitat layers
}
\examples{
proj_dir <- tempdir()
cache_dir <- fs::path(proj_dir, "cached-data")
get_emodnet_wfs_layers(layers = "be000142", cache_dir = cache_dir)
}
