% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h-emodnet.R
\name{get_emodnet_wfs_layers}
\alias{get_emodnet_wfs_layers}
\title{Access EMODnet EUNIS Seabed Habitat layers}
\usage{
get_emodnet_wfs_layers(
  layers,
  cache_dir = NULL,
  wfs_url = "https://ows.emodnet-seabedhabitats.eu/emodnet_open_maplibrary/wfs",
  crs = 4326,
  overwrite = FALSE
)
}
\arguments{
\item{layers}{character vector of EMODnet seabed habitat wfs layer names.}

\item{cache_dir}{path to the cached data directory. Defaults to the "cached-data'
folder in the root of the current project}

\item{wfs_url}{url to the EMODnet WFS service. Defaults to the EUNIS seabed habitat map wfs.}

\item{crs}{EPSG crs code to reproject to. Defaults to 4326}
}
\value{
an sf of requested features
}
\description{
Access EMODnet EUNIS Seabed Habitat layers
}
\examples{
proj_dir <- tempdir()
cache_dir <- fs::path(proj_dir, "cached-data")
get_emodnet_wfs_layers(layers = "be000142", cache_dir = cache_dir)
}
