% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis-api.R
\name{get_obis_recs}
\alias{get_obis_recs}
\title{Query OBIS for occurence data}
\usage{
get_obis_recs(
  sp_id,
  validate_sp_id = FALSE,
  fields = c("decimalLongitude", "decimalLatitude", "minimumDepthInMeters",
    "maximumDepthInMeters", "depth", "eventDate", "year", "month", "scientificName",
    "aphiaID"),
  geometry = NULL,
  ignore_failures = FALSE
)
}
\arguments{
\item{sp_id}{a valid WoRMS aphiaID}

\item{validate_sp_id}{whether to check for validity of aphiaID on OBIS}

\item{fields}{fields to return from OBIS query. Defaults to \code{c("decimalLongitude", "decimalLatitude","minimumDepthInMeters", "maximumDepthInMeters", "depth", "eventDate", "year", "month", "scientificName", "aphiaID")}}

\item{geometry}{an sf object. Queries will be faster if the
geometry is simple and restricted in space.}

\item{ignore_failures}{whether to through an error or return
NULL for queries that return no obis records.}
}
\value{
a tibble of OBIS occurence records
for \code{sp_id}.
}
\description{
Query OBIS for occurence data
}
\examples{
get_obis_recs(sp_id = 325567)
}
