% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis-api.R
\name{get_obis_recs}
\alias{get_obis_recs}
\title{Title}
\usage{
get_obis_recs(
  species_id,
  missing_check = FALSE,
  fields = c("decimalLongitude", "decimalLatitude", "minimumDepthInMeters",
    "maximumDepthInMeters", "depth", "eventDate", "year", "month", "scientificName",
    "aphiaID")
)
}
\arguments{
\item{species_id}{}

\item{missing_check}{}

\item{fields}{}
}
\value{

}
\description{
Title
}
