% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layer_info}
\alias{layer_info}
\title{Available habitate layer metadata}
\format{
A tibble of 34 variables and 1,459 observations
}
\usage{
layer_info
}
\description{
Metadata attributes of all available data layers
}
\keyword{datasets}
