% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis-to-h.R
\name{obis_match_habitat}
\alias{obis_match_habitat}
\title{Get OBIS occurences matched to layers of evironmental data}
\usage{
obis_match_habitat(
  sp_id,
  layers,
  fields = c("decimalLongitude", "decimalLatitude", "minimumDepthInMeters",
    "maximumDepthInMeters", "depth", "eventDate", "year", "month", "scientificName",
    "aphiaID"),
  check_match = FALSE,
  validate_sp_id = FALSE,
  geometry = NULL,
  ignore_failures = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{sp_id}{validates species Aphia ID}

\item{layers}{character vector of habitat layers to match}

\item{fields}{fields to return from OBIS query. Defaults to \code{c("decimalLongitude", "decimalLatitude","minimumDepthInMeters", "maximumDepthInMeters", "depth", "eventDate", "year", "month", "scientificName", "aphiaID")}}

\item{check_match}{whether to check validity of \code{sp_id}.}

\item{validate_sp_id}{whether to check for validity of aphiaID on OBIS}

\item{geometry}{an sf object. Queries will be faster if the
geometry is simple and restricted in space.}

\item{ignore_failures}{whether to through an error or return
NULL for queries that return no obis records.}

\item{overwrite}{whether to re-download and overwrite already downloaded layers}
}
\value{
an \code{sf} object, one row for each species occurence returned from obis.
Column include requested OBIS metadata through \code{fields} argument plus a column
for each layer environmental layer requested containing data retuned for each
occurence location.
}
\description{
function to get OBIS records for a given species and match to IAP and Bio-Oracle temperatures
if save_all_recs == TRUE, this full matched dataset will be saved before summarising
function returns a summary of temperature affinity of the species
}
\examples{
obis_match_habitat(sp_id = 325567, layers = c("BO_bathymean", 
                   "BO2_phosphateltmax_bdmean", "BO2_salinitymin_ss", 
                   "rock50cm", "april", "march"))
}
